Update tappingFeatures.R
Added a featuresFromColumn function that generalizes extracting features from a given column in a dataframe. A variable named processingFunction should be passed to it, to apply on each element of the selected column. For e.g. if the column is a JSON file location string, then ideally your processingFunction should read the JSON file and do something (pre-process/shape -> extract features etc.,) with it